TIP-451: create an abstractItemMediaWriter for CSV & XLSX writers
